mcp-server/requirements.txt
mcp
google-cloud-storage
google-generativeai
python-dotenv